+++ pkg/crmanager/types.go
@@ -144,6 +144,15 @@ type (
 		AllowVLANs             []string              `json:"allowVlans,omitempty"`
 		PersistenceMethods     []string              `json:"-"`
 	}
+
+	// ServiceAddress Service IP address definition (BIG-IP virtual-address).
+	ServiceAddress struct {
+		ArpEnabled         bool   `json:"arpEnabled,omitempty"`
+		ICMPEcho           string `json:"icmpEcho,omitempty"`
+		RouteAdvertisement string `json:"routeAdvertisement,omitempty"`
+		TrafficGroup       string `json:"trafficGroup,omitempty"`
+		SpanningEnabled    bool   `json:"spanningEnabled,omitempty"`
+	}
 	// Virtuals is slice of virtuals
 	Virtuals []Virtual
 
@@ -167,11 +176,12 @@ type (
 
 	// ResourceConfig is a Config for a single VirtualServer.
 	ResourceConfig struct {
-		MetaData metaData  `json:"-"`
-		Virtual  Virtual   `json:"virtual,omitempty"`
-		Pools    Pools     `json:"pools,omitempty"`
-		Policies Policies  `json:"policies,omitempty"`
-		Monitors []Monitor `json:"monitors,omitempty"`
+		MetaData       metaData         `json:"-"`
+		Virtual        Virtual          `json:"virtual,omitempty"`
+		Pools          Pools            `json:"pools,omitempty"`
+		Policies       Policies         `json:"policies,omitempty"`
+		Monitors       []Monitor        `json:"monitors,omitempty"`
+		ServiceAddress []ServiceAddress `json:"serviceAddress,omitempty"`
 	}
 	// ResourceConfigs is group of ResourceConfig
 	ResourceConfigs []*ResourceConfig
@@ -519,7 +529,7 @@ type (
 		TranslateServerAddress bool                 `json:"translateServerAddress,omitempty"`
 		TranslateServerPort    bool                 `json:"translateServerPort,omitempty"`
 		Class                  string               `json:"class,omitempty"`
-		VirtualAddresses       []string             `json:"virtualAddresses,omitempty"`
+		VirtualAddresses       []as3MultiTypeParam  `json:"virtualAddresses,omitempty"`
 		VirtualPort            int                  `json:"virtualPort,omitempty"`
 		SNAT                   as3MultiTypeParam    `json:"snat,omitempty"`
 		PolicyEndpoint         as3MultiTypeParam    `json:"policyEndpoint,omitempty"`
@@ -532,6 +542,17 @@ type (
 		ProfileL4              string               `json:"profileL4,omitempty"`
 		AllowVLANs             []as3ResourcePointer `json:"allowVlans,omitempty"`
 		PersistenceMethods     []string             `json:"persistenceMethods,omitempty"`
+	}
+
+	// as3ServiceAddress maps to VirtualAddress in AS3 Resources
+	as3ServiceAddress struct {
+		Class              string `json:"class,omitempty"`
+		VirtualAddress     string `json:"virtualAddress,omitempty"`
+		ArpEnabled         bool   `json:"arpEnabled,omitempty"`
+		ICMPEcho           string `json:"icmpEcho,omitempty"`
+		RouteAdvertisement string `json:"routeAdvertisement,omitempty"`
+		TrafficGroup       string `json:"trafficGroup,omitempty,omitempty"`
+		SpanningEnabled    bool   `json:"spanningEnabled,omitempty"`
 	}
 
 	// as3Monitor maps to the following in AS3 Resources